requirements.txt
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
courlan==0.2.3
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
geograpy3==0.1.24
htmldate==0.7.2
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
isodate==0.6.0
jellyfish==0.6.1
jieba3k==0.35.1
joblib==0.17.0
jusText==2.2.0
lxml==4.6.1
mongoengine==0.20.0
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
packaging==20.4
pandas==1.1.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pycountry==20.7.3
pylodstorage==0.0.18
pymongo==3.11.0
pypandoc==1.5
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
rdflib==5.0.0
readability-lxml==0.8.1
regex==2020.10.28
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
SPARQLWrapper==1.8.5
tinysegmenter==0.3
tldextract==3.0.2
toml==0.10.2
tqdm==4.51.0
trafilatura==0.5.2
twarc==1.11.0
Unidecode==1.1.1
urllib3==1.25.11
us==2.0.2
zipp==3.4.0
